backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
watchdog
pdfplumber
python-docx
sentence-transformers
numpy
openai
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
